// cape-playground/src/app/cape-sdk-angular/model/service-link/serviceLinkRecordDoubleSigned.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ServiceLinkRecordPayload } from './serviceLinkRecordPayload';
import { ServiceLinkRecordSignature } from './serviceLinkRecordSignature';
import { ServiceLinkStatusRecordSigned } from './serviceLinkStatusRecordSigned';


export interface ServiceLinkRecordDoubleSigned { 
    payload: ServiceLinkRecordPayload;
    signatures: Array<ServiceLinkRecordSignature>;
    serviceLinkStatuses?: Array<ServiceLinkStatusRecordSigned>;
}