// cape-playground/src/app/cape-sdk-angular/model/service-link/serviceLinkStatusRecordPayload.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServiceLinkStatusRecordPayload {
  iat: string;
  version: string;
  record_id: string;
  surrogate_id: string;
  slr_id: string;
  sl_status: SlStatusEnum;
  prev_record_id: string;
}
export enum SlStatusEnum{
  Active = 'Active',
  Removed = 'Removed'
}