// cape-playground/src/app/cape-sdk-angular/model/service-link/serviceLinkStatusRecordSigned.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ServiceLinkStatusRecordPayload } from './serviceLinkStatusRecordPayload';
import { JWSHeader } from './jWSHeader';


export interface ServiceLinkStatusRecordSigned { 
    header: JWSHeader;
    payload: ServiceLinkStatusRecordPayload;
    signature: object;
    _protected: object;
}