// cape-playground/src/app/cape-sdk-angular/model/consent/consentRecordSigned.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConsentRecordPayload } from './consentRecordPayload';
import { ConsentStatusRecordSigned } from './consentStatusRecordSigned';


export interface ConsentRecordSigned { 
    payload: ConsentRecordPayload;
    signature: object;
    consentStatusList?: Array<ConsentStatusRecordSigned>;
    _protected: object;
}