// cape-playground/src/app/cape-sdk-angular/model/consent/consentStatusRecordPayload.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SinkUsageRules } from './sinkUsageRules';
import { ResourceSet } from './resourceSet';


export interface ConsentStatusRecordPayload {
  record_id: string;
  version: string;
  surrogate_id: string;
  cr_id: string;
  consent_status: ConsentStatusEnum;
  consent_resource_set?: ResourceSet;
  consent_usage_rules?: SinkUsageRules;
  iat: string;
  prev_record_id: string;
}

export enum ConsentStatusEnum {
  Active = 'Active',
  Disabled = 'Disabled',
  Withdrawn = 'Withdrawn'
}