// cape-playground/src/app/cape-sdk-angular/model/consent/consentStatusRecordSigned.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConsentStatusRecordPayload } from './consentStatusRecordPayload';


export interface ConsentStatusRecordSigned { 
    payload: ConsentStatusRecordPayload;
    signature: object;
    _protected: object;
}