// cape-playground/src/app/cape-sdk-angular/model/service-link/serviceEntry.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProcessingBasis } from '../processingBasis';
import { HumanReadableDescription } from '../humanReadableDescription';
import { HasCompetentAuthority } from './hasCompetentAuthority';
import { ServiceInstance } from './serviceInstance';
import { IsDescribedAt } from './isDescribedAt';


export interface ServiceEntry {
  serviceId?: string;
  name?: string;
  identifier?: string;
  issued?: string;
  createdByUserId?: string;
  serviceDescriptionVersion?: string;
  serviceIconUrl?: string;
  status?: string;
  supportedProfiles?: Array<SupportedProfilesEnum>;
  spatial?: string;
  lifeEvent?: string;
  businessEvent?: string;
  sector?: Array<string>;
  type?: Array<string>;
  thematicArea?: Array<string>;
  language?: Array<string>;
  hasCompetentAuthority?: HasCompetentAuthority;
  humanReadableDescription?: Array<HumanReadableDescription>;
  serviceInstance?: ServiceInstance;
  isDescribedAt?: Array<IsDescribedAt>;
  processingBases?: Array<ProcessingBasis>;
  role?: RoleEnum;
}
export enum SupportedProfilesEnum {
  Consenting = 'Consenting',
  _3rdPartyReUse = '3rd party re-use',
  Notification = 'Notification',
  Objection = 'Objection'
}

export enum RoleEnum {
  Sink = 'Sink',
  Source = 'Source'
}